JRUBY-2643: Fix JRuby newFixnum issue causing NSMEs w/ 1.1.2
git-svn-id: svn+ssh://rubyforge.org/var/svn/jruby-extras/trunk/activerecord-jdbc@1026 8ba958d5-0c1a-0410-94a6-a65dfc1b28a6

+++ src/java/jdbc_adapter/JdbcAdapterInternalService.java
@@ -614,7 +614,7 @@ public class JdbcAdapterInternalService implements BasicLibraryService {
                 Statement stmt = null;
                 try {
                     stmt = c.createStatement();
-                    return recv.getRuntime().newFixnum(stmt.executeUpdate(rubyApi.convertToRubyString(sql).getUnicodeValue()));
+                    return recv.getRuntime().newFixnum((long)stmt.executeUpdate(rubyApi.convertToRubyString(sql).getUnicodeValue()));
                 } finally {
                     if (null != stmt) {
                         try {
